refactor: pass both image and tool_input as parameters to capture
Allow tools to decide how to handle capturing results based on just an image vs some other input such as an SBOM file.

For grype and syft it didn't matter because both types of input were handled the same way; however, for other tools (like anchorectl), the commands will be different and both the image name and path to the sbom will be needed.

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>