chore: change dependabot to auto-approve only
When dependabot PRs are automatically merged, the resulting commit
doesn't run workflows on main if the default GitHub token is used.
Rather than generate a more potent token, or live with some commits not
running workflows on main, scale back dependabot automation to only
approve the PRs but not merge them. This still reduces toil by saving
several clicks in the UI.

Signed-off-by: Will Murphy <will.murphy@anchore.com>